nix: don't use nix flake
I originally wanted to use the nix flake to allow users to take
advantage of the latest changes. Just so happened that nixpkgs was
recently updated with a new version around the same time, and this
seems to be the root cause of a bunch of ci failures, so we will just
pull the new version from nixpkgs like before.